Resolve dependency circle
